requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.37.0
peft>=0.7.0
datasets>=2.16.0
accelerate>=0.25.0

# Embeddings
sentence-transformers>=2.3.0

# Evaluation
scikit-learn>=1.3.0

# Config
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
numpy>=1.24.0
langdetect>=1.0.9

# Optional
# wandb>=0.16.0
